+++ src/Image/Palette/Color/Gray.php
@@ -96,7 +96,8 @@ final class Gray implements ColorInterface
     public function dissolve($alpha)
     {
         return $this->palette->color(
-            array($this->gray), $this->alpha + $alpha
+            array($this->gray),
+            min(max((int) round($this->alpha + $alpha), 0), 100)
         );
     }