+++ tests/Imagine/Test/Image/Palette/Color/AbstractColorTest.php
@@ -54,10 +54,17 @@ abstract class AbstractColorTest extends ImagineTestCase
         $alpha = $color->getAlpha();
         $signature = (string) $color;
 
-        $color = $color->dissolve(2);
+        $color2 = $color->dissolve(2);
+        $this->assertEquals(2 + $alpha, $color2->getAlpha());
+        $this->assertEquals($signature, (string) $color2);
 
-        $this->assertEquals(2 + $alpha, $color->getAlpha());
-        $this->assertEquals($signature, (string) $color);
+        $color2 = $color->dissolve(200);
+        $this->assertEquals(100, $color2->getAlpha());
+        $this->assertEquals($signature, (string) $color2);
+
+        $color2 = $color->dissolve(-200);
+        $this->assertEquals(0, $color2->getAlpha());
+        $this->assertEquals($signature, (string) $color2);
     }
 
     public function testLighten()